requirements.txt
# CAD Compliance RAG (runtime)
# Standard library only.

# one-time only
python-dotenv>=1.0.0
mistralai>=1.0.0